// src/server/model/iOtherChargesLine.ts
/**
 * 
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { ICostCenterLine } from './iCostCenterLine';

export interface IOtherChargesLine { 
    value?: number;
    costCenterLines?: Array<ICostCenterLine>;
    amount?: number;
    chargesId?: number;
    chargesName?: string;
    id?: number;
    jacksontype?: string;
}